+++ src/Alfil.h
@@ -9,6 +9,6 @@ public:
 	void Dibuja();
 	void mueve(Vector2D);
 
-	Alfil(bool _color, float _x, float _y);
+	Alfil(bool _color, int _x, int _y);
 
 };
