+++ src/Peon.h
@@ -8,5 +8,5 @@ private:
 public:
 	void Dibuja();
 	void mueve(Vector2D);
-	Peon(bool _color, float _x, float _y);
+	Peon(bool _color, int _x, int _y);
 };
