+++ src/Rey.h
@@ -9,6 +9,6 @@ public:
 	void Dibuja();
 	void mueve(Vector2D);
 
-	Rey(bool _color, float _x, float _y);
+	Rey(bool _color, int _x, int _y);
 
 };
